--- frontend/sprios_ios/sprios_ios/Controllers/signUp/AddUsernameController.swift
@@ -1,15 +1,0 @@
-//
-//  AddUsernameController.swift
-//  sprios_ios
-//
-//  Created by 지윤 on 2022/12/10.
-//
-
-import UIKit
-class AddUsernameController: UIViewController {
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-
-}
