+++ frontend/sprios_ios/sprios_ios/JoinViewController.swift
@@ -1,0 +1,30 @@
+//
+//  JoinViewController.swift
+//  sprios_ios
+//
+//  Created by 이정동 on 2022/11/28.
+//
+
+import UIKit
+
+class JoinViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        view.backgroundColor = .systemBackground
+        // Do any additional setup after loading the view.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
