+++ frontend/sprios_ios/sprios_ios/SceneDelegate.swift
@@ -17,20 +17,6 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // If using a storyboard, the `window` property will automatically be initialized and attached to the scene.
         // This delegate does not imply the connecting scene or session are new (see `application:configurationForConnectingSceneSession` instead).
         guard let _ = (scene as? UIWindowScene) else { return }
-        
-        let storyboard = UIStoryboard(name: "Main", bundle: nil)
-        
-        let loginNavController = storyboard.instantiateViewController(withIdentifier: "LoginNavigationController")
-        
-        window?.rootViewController = loginNavController
-    }
-    
-    // RootViewController 변경
-    func changeRootViewController(_ vc: UIViewController, animated: Bool = true) {
-        guard let window = self.window else { return }
-        
-        // change the root view controller to your specific view controller
-        window.rootViewController = vc
     }
 
     func sceneDidDisconnect(_ scene: UIScene) {
