--- frontend/sprios_ios/sprios_ios/Views/Home/FeedCell.swift
@@ -1,23 +1,0 @@
-//
-//  FeedCell.swift
-//  sprios_ios
-//
-//  Created by 이정동 on 2022/11/30.
-//
-
-import UIKit
-
-class FeedCell: UITableViewCell {
-
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        // Initialization code
-    }
-
-    override func setSelected(_ selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-
-        // Configure the view for the selected state
-    }
-    
-}
