--- frontend/sprios_ios/sprios_ios/Views/PhotoCell.swift
@@ -1,17 +1,0 @@
-//
-//  PhotoCell.swift
-//  sprios_ios
-//
-//  Created by 이정동 on 2022/12/07.
-//
-
-import UIKit
-
-class PhotoCell: UICollectionViewCell {
-    
-    @IBOutlet weak var photo: UIImageView!
-    
-    override func prepareForReuse() {
-        self.photo.image = nil
-    }
-}
